requirements.txt
fastapi~=0.103.2
pydantic~=2.4.2
psycopg2-binary~=2.8.4
uvicorn